Add test for read_key and separate the other domain objects
